Move TestNode2ContainerMap to org.apache.hadoop.hdds.scm.container package. Make Node2ContainerMap package-private
